Add files via upload
This folder has two files, one python and the other one a jupyter notebook. My anaconda and jupyter had some internal import error which couldn't be rectified but the code worked fine in python and pycharm ide. The CSV file has been submitted and some screenshot would also be there to show the console of this.